+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Implement `xsd:normalizedString` data type codec
 - Implement `xsd:Name` data type codec
 - Implement `xsd:token` data type codec
+- Implement `xsd:int` data type codec